requirements.txt
typer==0.12.5
rich==13.9.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
pyyaml==6.0.2
nltk==3.9.1

# Optional (used if available; not required for basic run)
# langchain==0.2.12
# langchain-community==0.2.11
# sentence-transformers==3.0.1
# crewai==0.28.8
streamlit==1.38.0
PyPDF2==3.0.1
plotly==5.23.0

